+++ biobb_structure_checking/constants.py
@@ -257,6 +257,8 @@ DIALOGS.add_entry('chains', 'Checks and selects chains')
 DIALOGS.add_option('chains', '--select', 'select',\
     'Chains (All | protein | na | dna | rna | Chain list comma separated)')
 DIALOGS.add_option('chains', '--rename', 'rename','Rename unlabelled chains (auto | label)')
+DIALOGS.add_option('chains', '--renumber', 'renumber','Renumber residues (auto | [A|*:]ini0[-fin0]=[B|*:]ini1[-fin1])')
+DIALOGS.add_option('chains', '--allow_merge', 'allow_merge','Allow merge several chains in a single one', 'bool')
 
 DIALOGS.add_entry('altloc', 'Checks and selects alternative locations')
 DIALOGS.add_option('altloc', '--select', 'select', \