+++ biobb_structure_checking/structure_manager.py
@@ -716,6 +716,21 @@ class StructureManager:
         self.update_internals()
         self.modified = True
 
+    def rename_empty_chain_label(self, new_label):
+        '''Add labels to unlabelled chains'''
+        result = self.chains_data.rename_empty_chain_label(new_label)
+        self.update_internals()
+        self.modified = True
+        return result
+
+    def renumber_chain_residues(self, renum_str, allow_merge=False):
+        ''' Allow to relabel chains and residues'''
+        result = self.chains_data.renumber(renum_str, allow_merge=allow_merge)
+        if result:
+            self.update_internals()
+            self.modified = True
+        return result
+
     def select_altloc_residue(self, res: Residue, to_fix: Mapping[str, Union[str, Atom]]) -> None:
         """ Selects one alternative conformation when altloc exists. Selection is
             done on the occupancy basis or from the conformation id.
@@ -1291,7 +1306,7 @@ class StructureManager:
             res[amide_res[res_type][0]],
             res[amide_res[res_type][1]]
         )
-        
+
     def _amide_score(self, matr):
         score = 0.
         for amide_res in sorted(matr):
@@ -1307,13 +1322,13 @@ class StructureManager:
                         (d and m1 and not m2):
                             score += 1/matr[amide_res]['cnts'][atm][atm_cnt]
         return score
-        
+
     def _amide_cluster(self, to_fix):
         cluster = {}
         for res in to_fix:
             cluster[res] = set()
             cluster[res].add(res)
-        
+
         for r_pair in self.rr_dist:
             res1, res2 = r_pair[0:2]
             if res1 in to_fix and res2 in to_fix:
@@ -1329,7 +1344,7 @@ class StructureManager:
                     if res2 in cluster:
                         del cluster[res2]
         return cluster
-    
+
     def _is_amide_atom(self, amide_res, atm):
         res = atm.get_parent()
         return res.get_resname() in amide_res and atm.id in amide_res[res.get_resname()]
@@ -1347,7 +1362,7 @@ class StructureManager:
                 at1, at2, dist2 = cnt
                 if at1.serial_number > at2.serial_number:
                     at2, at1, dist2 = cnt
-        matr = {}        
+        matr = {}
         for res_pair in c_list['polar']:
             for cnt in c_list['polar'][res_pair]:
                 at1, at2, dist2 = cnt